refactor: startingpoint -> departure 패키지 명 변경
